#4346 Move aggregate step handler to upper level
