+++ app/integration/runtime/src/main/java/io/syndesis/integration/runtime/IntegrationRouteBuilder.java
@@ -36,6 +36,7 @@ import io.syndesis.common.model.integration.StepKind;
 import io.syndesis.common.util.Json;
 import io.syndesis.common.util.KeyGenerator;
 import io.syndesis.integration.runtime.capture.OutMessageCaptureProcessor;
+import io.syndesis.integration.runtime.handlers.AggregateStepHandler;
 import io.syndesis.integration.runtime.handlers.ConnectorStepHandler;
 import io.syndesis.integration.runtime.handlers.DataMapperStepHandler;
 import io.syndesis.integration.runtime.handlers.EndpointStepHandler;
@@ -45,7 +46,7 @@ import io.syndesis.integration.runtime.handlers.HeadersStepHandler;
 import io.syndesis.integration.runtime.handlers.LogStepHandler;
 import io.syndesis.integration.runtime.handlers.RuleFilterStepHandler;
 import io.syndesis.integration.runtime.handlers.SimpleEndpointStepHandler;
-import io.syndesis.integration.runtime.handlers.SplitAggregateStepHandler;
+import io.syndesis.integration.runtime.handlers.SplitStepHandler;
 import io.syndesis.integration.runtime.handlers.TemplateStepHandler;
 import io.syndesis.integration.runtime.logging.ActivityTracker;
 import io.syndesis.integration.runtime.logging.ActivityTrackingPolicy;
@@ -93,8 +94,8 @@ public class IntegrationRouteBuilder extends RouteBuilder {
         this.stepHandlerList.add(new ExpressionFilterStepHandler());
         this.stepHandlerList.add(new RuleFilterStepHandler());
         this.stepHandlerList.add(new ExtensionStepHandler());
-        this.stepHandlerList.add(new SplitAggregateStepHandler());
-        this.stepHandlerList.add(new SplitAggregateStepHandler.AggregateStepHandler());
+        this.stepHandlerList.add(new SplitStepHandler());
+        this.stepHandlerList.add(new AggregateStepHandler());
         this.stepHandlerList.add(new LogStepHandler());
         this.stepHandlerList.add(new HeadersStepHandler());
         this.stepHandlerList.add(new TemplateStepHandler());
