+++ app/integration/runtime/src/main/java/io/syndesis/integration/runtime/handlers/AggregateStepHandler.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package io.syndesis.integration.runtime.handlers;
 
 import javax.annotation.concurrent.Immutable;
@@ -31,67 +32,21 @@ import io.syndesis.common.model.integration.StepKind;
 import io.syndesis.integration.runtime.IntegrationRouteBuilder;
 import io.syndesis.integration.runtime.IntegrationStepHandler;
 import org.apache.camel.Exchange;
-import org.apache.camel.Expression;
-import org.apache.camel.builder.Builder;
 import org.apache.camel.model.ProcessorDefinition;
 import org.apache.camel.processor.aggregate.AggregationStrategy;
 import org.apache.camel.processor.aggregate.GroupedBodyAggregationStrategy;
 import org.apache.camel.processor.aggregate.UseLatestAggregationStrategy;
 import org.apache.camel.processor.aggregate.UseOriginalAggregationStrategy;
-import org.apache.camel.spi.Language;
-import org.apache.camel.util.ObjectHelper;
 
-public class SplitAggregateStepHandler implements IntegrationStepHandler {
+public class AggregateStepHandler implements IntegrationStepHandler {
     @Override
     public boolean canHandle(Step step) {
-        return StepKind.split == step.getStepKind();
+        return StepKind.aggregate == step.getStepKind();
     }
 
-    @SuppressWarnings({"PMD.AvoidReassigningParameters", "PMD.AvoidDeeplyNestedIfStmts"})
     @Override
     public Optional<ProcessorDefinition<?>> handle(Step step, ProcessorDefinition<?> route, IntegrationRouteBuilder builder, String flowIndex, String stepIndex) {
-        ObjectHelper.notNull(route, "route");
-
-        String languageName = step.getConfiguredProperties().get("language");
-        String expressionDefinition = step.getConfiguredProperties().get("expression");
-        AggregationOption aggregation = Optional.ofNullable(step.getConfiguredProperties().get("aggregationStrategy"))
-                                                        .map(AggregationOption::valueOf)
-                                                        .orElse(AggregationOption.body);
-
-        if (ObjectHelper.isEmpty(languageName) && ObjectHelper.isEmpty(expressionDefinition)) {
-            route = route.split(Builder.body()).aggregationStrategy(aggregation.getStrategy(step.getConfiguredProperties()));
-        } else if (ObjectHelper.isNotEmpty(expressionDefinition)) {
-
-            if (ObjectHelper.isEmpty(languageName)) {
-                languageName = "simple";
-            }
-
-            // A small hack until https://issues.apache.org/jira/browse/CAMEL-12079
-            // gets fixed so we can support the 'bean::method' annotation as done by
-            // Function step definition
-            if ("bean".equals(languageName) && expressionDefinition.contains("::")) {
-                expressionDefinition = expressionDefinition.replace("::", "?method=");
-            }
-
-            final Language language = builder.getContext().resolveLanguage(languageName);
-            final Expression expression = language.createExpression(expressionDefinition);
-
-            route = route.split(expression).aggregationStrategy(aggregation.getStrategy(step.getConfiguredProperties()));
-        }
-
         return Optional.of(route);
-    }
-
-    public static class AggregateStepHandler implements IntegrationStepHandler {
-        @Override
-        public boolean canHandle(Step step) {
-            return StepKind.aggregate == step.getStepKind();
-        }
-
-        @Override
-        public Optional<ProcessorDefinition<?>> handle(Step step, ProcessorDefinition<?> route, IntegrationRouteBuilder builder, String flowIndex, String stepIndex) {
-            return Optional.empty();
-        }
     }
 
     @SuppressWarnings("unchecked")
@@ -99,7 +54,7 @@ public class SplitAggregateStepHandler implements IntegrationStepHandler {
         body(GroupedBodyAggregationStrategy::new),
         latest(UseLatestAggregationStrategy::new),
         original(() -> new UseOriginalAggregationStrategy(null, true)),
-        script(ScriptAggregationStrategy::new, (strategy, stepProperties) -> {
+        script(AggregateStepHandler.ScriptAggregationStrategy::new, (strategy, stepProperties) -> {
             Optional.ofNullable(stepProperties.get("aggregationScriptLanguage")).ifPresent(strategy::setLanguage);
             Optional.ofNullable(stepProperties.get("aggregationScript")).ifPresent(strategy::setScript);
             return strategy;
